+++ public/app/features/dashboard-scene/panel-edit/PanelEditNext/QueryEditor/testUtils.tsx
@@ -2,7 +2,7 @@ import { render } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { ReactElement } from 'react';
 
-import { DataSourceInstanceSettings, PluginType } from '@grafana/data';
+import { DataSourceInstanceSettings, getDefaultTimeRange, LoadingState, PluginType } from '@grafana/data';
 import { VizPanel } from '@grafana/scenes';
 import { DataQuery } from '@grafana/schema';
 import { QueryGroupOptions } from 'app/types/query';
@@ -10,14 +10,14 @@ import { QueryGroupOptions } from 'app/types/query';
 import { QueryEditorType } from '../constants';
 
 import {
+  AlertingState,
   DatasourceState,
   PanelState,
   QueryEditorActions,
+  QueryEditorProvider,
   QueryEditorUIState,
   QueryOptionsState,
   QueryRunnerState,
-  QueryEditorProvider,
-  AlertingState,
 } from './QueryEditorContext';
 import { Transformation } from './types';
 
@@ -159,7 +159,11 @@ export function renderWithQueryEditorProvider(children: ReactElement, options: C
 
   const defaultQrState: QueryRunnerState = {
     queries,
-    data: undefined,
+    data: {
+      state: LoadingState.Done,
+      series: [],
+      timeRange: getDefaultTimeRange(),
+    },
     isLoading: false,
     queryError: undefined,
     ...qrState,
